// src/erp-ui/projects/api/src/lib/model/booking-summary-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaxAreaSummaryDto } from './tax-area-summary-dto';
import { MonthSummaryDto } from './month-summary-dto';


export interface BookingSummaryDto { 
    year: number;
    income: number;
    expense: number;
    balance: number;
    byTaxArea: Array<TaxAreaSummaryDto>;
    byMonth: Array<MonthSummaryDto>;
}